split away cosmos WasmClient
