+++ packages/strategist-utils/src/cosmos/mod.rs
@@ -1,2 +1,3 @@
 pub mod base_client;
 pub mod errors;
+pub mod wasm_client;
